moved shared visit view code to org.smartregister.chw.anc.fragment.BaseHomeVisitFragment
